refactor: remove owner field from base book vh
